+++ src/app/applications/addcloudapplication.component.html
@@ -391,7 +391,7 @@
                             cloud
                             usage policy.</p>
                     </div>
-                    <button (click)="filterEnteredData(f); submitModal.show();"
+                    <button (click)="extension_request=false;filterEnteredData(f); submitModal.show();"
                             class="btn btn-sm btn-primary" [disabled]="f.invalid">
                         <i class="fa fa-dot-circle-o"></i>
                         Submit Application
