+++ src/app/applications/addcloudapplication.component.ts
@@ -5,8 +5,8 @@ import {ApplicationsService} from '../api-connector/applications.service'
 import {FlavorService} from '../api-connector/flavor.service';
 import {Flavor} from '../virtualmachines/virtualmachinemodels/flavor';
 import {FlavorType} from '../virtualmachines/virtualmachinemodels/flavorType';
-import {AbstractBaseClasse} from '../shared/shared_modules/baseClass/abstract-base-class';
 import {environment} from '../../environments/environment';
+import {ApplicationBaseClass} from 'app/shared/shared_modules/baseClass/application-base-class';
 
 /**
  * This components provides the functions to create a new Cloud Application.
@@ -18,7 +18,7 @@ import {environment} from '../../environments/environment';
     styleUrls: ['addcloudapplication.component.css']
 })
 
-export class AddcloudapplicationComponent extends AbstractBaseClasse {
+export class AddcloudapplicationComponent extends ApplicationBaseClass {
 
     /**
      * If it is in production or dev mode.
@@ -30,14 +30,6 @@ export class AddcloudapplicationComponent extends AbstractBaseClasse {
      * List of all collapse booleans.
      */
     public collapseList: boolean[];
-
-    public project_application_report_allowed: boolean = false;
-
-    /**
-     * If shortname is valid.
-     * @type {boolean}
-     */
-    public wronginput: boolean = false;
 
     /**
      * Contains errors recieved when submitting an application.
@@ -57,29 +49,7 @@ export class AddcloudapplicationComponent extends AbstractBaseClasse {
      */
     public typeList: FlavorType[];
 
-    /**
-     * Total number of cores.
-     * @type {number}
-     */
-    public totalNumberOfCores: number = 0;
-    /**
-     * Total number of ram.
-     * @type {number}
-     */
-    public totalRAM: number = 0;
-    /**
-     * Values to confirm.
-     */
-    public valuesToConfirm: string[];
-    /**
-     *
-     */
-    public constantStrings: Object;
 
-    /**
-     * Name of the project.
-     */
-    public projectName: string;
     public acknowledgeModalTitle: string = 'Acknowledge';
     public acknowledgeModalType: string = 'info';
 
@@ -95,113 +65,11 @@ export class AddcloudapplicationComponent extends AbstractBaseClasse {
      * @param {ApplicationsService} applicationsservice
      * @param {FlavorService} flavorservice
      */
-    constructor(private applicationsservice: ApplicationsService, private flavorservice: FlavorService) {
-        super();
+    constructor(applicationsservice: ApplicationsService, private flavorservice: FlavorService) {
+        super(null, null, applicationsservice, null);
         this.getListOfFlavors();
         this.getListOfTypes();
 
-    }
-
-    /**
-     * Uses the data from the application form to fill the confirmation-modal with information.
-     * @param form the application form with corresponding data
-     */
-    filterEnteredData(form: NgForm): void {
-        this.generateConstants();
-        this.totalNumberOfCores = 0;
-        this.totalRAM = 0;
-        this.valuesToConfirm = [];
-        for (const key in form.controls) {
-            if (form.controls[key].value) {
-                if (key === 'project_application_name') {
-                    this.projectName = form.controls[key].value;
-                    if (this.projectName.length > 50) {
-                        this.projectName = `${this.projectName.substring(0, 50)}...`;
-                    }
-                }
-                if (key in this.constantStrings) {
-                    this.valuesToConfirm.push(this.matchString(key.toString(), form.controls[key].value.toString()));
-
-                    const flavor: Flavor = this.isKeyFlavor(key.toString());
-                    if (flavor != null) {
-                        this.totalNumberOfCores = this.totalNumberOfCores + (flavor.vcpus * form.controls[key].value);
-                        const ram: number = flavor.ram * form.controls[key].value;
-                        this.totalRAM = this.totalRAM + ram
-                    }
-                }
-            }
-        }
-        if (!this.project_application_report_allowed) {
-            this.valuesToConfirm.push('Dissemination allowed: No')
-        }
-
-    }
-
-    /**
-     * Fills the array constantStrings with values dependent of keys which are used to indicate inputs from the application-form
-     */
-    generateConstants(): void {
-        this.constantStrings = [];
-        this.constantStrings['project_application_lifetime'] = 'Lifetime of your project: ';
-        this.constantStrings['project_application_volume_counter'] = 'Number of volumes for additional storage: ';
-        this.constantStrings['project_application_object_storage'] = 'Additional object storage: ';
-        this.constantStrings['project_application_volume_limit'] = 'Additional storage space for your VMs: ';
-        this.constantStrings['project_application_institute'] = 'Your institute: ';
-        this.constantStrings['project_application_workgroup'] = 'Your Workgroup: ';
-        this.constantStrings['project_application_report_allowed'] = 'Dissemination allowed: ';
-
-        for (const key in this.flavorList) {
-            if (key in this.flavorList) {
-                this.constantStrings[`project_application_${this.flavorList[key].name}`] =
-                    `Number of VMs of type  ${this.flavorList[key].name}: `;
-            }
-        }
-    }
-
-    isKeyFlavor(key: string): Flavor {
-        for (const fkey in this.flavorList) {
-            if (fkey in this.flavorList) {
-                if (this.flavorList[fkey].name === key.substring(20)) {
-                    return this.flavorList[fkey];
-                }
-            }
-        }
-
-        return null;
-
-    }
-
-    /**
-     * This function concatenates a given key combined with a given value to a string
-     * which is used on the confirmation-modal.
-     * @param key the key to access a string in the array constantStrings
-     * @param val the value that is concatenated with the string from the array and an optional addition (depending on the key)
-     * @returns the concatenated string for the confirmation-modal
-     */
-    matchString(key: string, val: string): string {
-        if (key in this.constantStrings) {
-            switch (key) {
-                case 'project_application_lifetime': {
-                    return (`${this.constantStrings[key]}${val} months`);
-                }
-                case ('project_application_volume_limit'): {
-                    return (`${this.constantStrings[key]}${val} GB`);
-                }
-                case 'project_application_object_storage': {
-                    return (`${this.constantStrings[key]}${val}  GB`);
-                }
-                case 'project_application_report_allowed': {
-                    if (val) {
-                        return (`${this.constantStrings[key]}${val} Yes`);
-                    } else {
-                        return (`${this.constantStrings[key]} No`);
-                    }
-                }
-                default: {
-                    return (this.constantStrings[key] + val);
-                }
-            }
-        }
     }
 
     /**
@@ -323,11 +191,4 @@ export class AddcloudapplicationComponent extends AbstractBaseClasse {
 
     }
 
-    /**
-     * Check if shortname is valid.
-     * @param {string} shortname
-     */
-    public checkShortname(shortname: string): void {
-        this.wronginput = !/^[a-zA-Z0-9\s]*$/.test(shortname);
-    }
 }
