+++ src/app/applications/addsinglevm.component.html
@@ -346,7 +346,7 @@
                             cloud
                             usage policy.</p>
                     </div>
-                    <button (click)="acknowledgemodal.show()"
+                    <button (click)="extension_request=false;filterEnteredData(f); submitModal.show();"
                             [disabled]="f.invalid"
                             class="btn btn-sm btn-primary"><i
                             class="fa fa-dot-circle-o"></i> Submit
@@ -437,3 +437,60 @@
         </div><!-- /.modal-content -->
     </div><!-- /.modal-dialog -->
 </div><!-- /.modal -->
+
+
+<div bsModal #submitModal="bs-modal" class="modal fade" tabindex="-1"
+     role="dialog" aria-labelledby="Label"
+     aria-hidden="true">
+    <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h4 class="modal-title">{{projectName}}</h4>
+                <button type="button" class="close" style="cursor: pointer;"
+                        (click)="submitModal.hide()"
+                        data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+
+            </div>
+            <div class="modal-body">
+                <div class="alert alert-info">
+                    <strong>Please check the data you have entered and confirm your application
+                        below.</strong><br>
+                    If the entered data mismatches with your requirements cancel this confirmation
+                    and take changes.
+                </div>
+                <div class="table-responsive">
+                    <table class="table table-striped table-responsive">
+                        <tr *ngFor="let str of valuesToConfirm">
+                            <td>{{str}}</td>
+                        </tr>
+                        <tr>
+                            <td><strong>Number of cores in total:
+                                {{totalNumberOfCores.toString()}}</strong></td>
+                        </tr>
+                        <tr>
+                            <td><strong>Value of RAM in total: {{totalRAM.toString()}} GBs</strong>
+                            </td>
+                        </tr>
+                    </table>
+                </div>
+            </div>
+
+
+            <div class="modal-footer">
+
+
+                <button class="btn btn-success col-md4" type="button"
+                        (click)="acknowledgemodal.show(); submitModal.hide();">Confirm Application
+                </button>
+                <button class="btn  btn-primary col-md-4" type="reset"
+                        (click)="submitModal.hide();">Cancel
+                </button>
+
+            </div>
+        </div><!-- /.modal-content -->L
+    </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
+
+
