+++ src/app/applications/applications.component.html
@@ -270,7 +270,7 @@
                                                     {{computeCenter.Name}}
                                                 </option>
                                             </select>
-                                            <span  *ngIf="application.Status != application_states.SUBMITTED">{{application.ComputeCenter?.Name}}</span>
+                                            <span *ngIf="application.Status != application_states.SUBMITTED">{{application.ComputeCenter?.Name}}</span>
                                         </td>
 
 
@@ -1274,7 +1274,7 @@
                         <label class="col-md-4 form-control-label"
                                for="text-input"><strong>Comment</strong></label>
                         <div class="col-md-8">
-                <textarea type="text" id="id_project_application_renewal__comment"
+                <textarea type="text" id="id_project_application_renewal_comment"
                           name="project_application_renewal_comment" class="form-control"
                           maxlength="8192"
                           placeholder='' [ngModel]="selectedApplication?.Comment"></textarea>
@@ -1302,9 +1302,9 @@
                     <div class="label">
                         By submitting this application you agree to the cloud usage policy.
                     </div>
-                    <button type="button" class="btn  btn-primary" type="submit" name="_save"
+                    <button type="button" class="btn  btn-primary"
                             form="application_form"
-                            (click)="extensionModal.hide();resultModal.show()"
+                            (click)="extension_request=true;filterEnteredData(f); submitModal.show();"
                             [disabled]="f.invalid"><i
                             class="fa fa-dot-circle-o">
 
@@ -1322,6 +1322,60 @@
         </div><!-- /.modal-dialog -->
     </div><!-- /.modal -->
 </div>
+
+<div bsModal #submitModal="bs-modal" class="modal fade" tabindex="-1"
+     role="dialog" aria-labelledby="Label"
+     aria-hidden="true">
+    <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h4 class="modal-title">{{projectName}}</h4>
+                <button type="button" class="close" style="cursor: pointer;"
+                        (click)="submitModal.hide()"
+                        data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+
+            </div>
+            <div class="modal-body">
+                <div class="alert alert-info">
+                    <strong>Please check the data you have entered and confirm your application
+                        below.</strong><br>
+                    If the entered data mismatches with your requirements cancel this confirmation
+                    and take changes.
+                </div>
+                <div class="table-responsive">
+                    <table class="table table-striped table-responsive">
+                        <tr *ngFor="let str of valuesToConfirm">
+                            <td>{{str}}</td>
+                        </tr>
+                        <tr>
+                            <td><strong>Number of cores in total:
+                                {{totalNumberOfCores.toString()}}</strong></td>
+                        </tr>
+                        <tr>
+                            <td><strong>Value of RAM in total: {{totalRAM.toString()}} GBs</strong>
+                            </td>
+                        </tr>
+                    </table>
+                </div>
+            </div>
+
+
+            <div class="modal-footer">
+
+
+                <button class="btn btn-success col-md4" type="button"
+                        (click)="resultModal.show(); extensionModal.hide(); submitModal.hide();">Confirm Application
+                </button>
+                <button class="btn  btn-primary col-md-4" type="reset"
+                        (click)="submitModal.hide();">Cancel
+                </button>
+
+            </div>
+        </div><!-- /.modal-content -->L
+    </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
 
 
 <div bsModal #resultModal="bs-modal" class="modal fade" tabindex="-1" role="dialog"
